Moved ListenForEffectChanges to Refresh
This fixes a bug where it was impossible to get the TextEffect to listen
for effect changes if they were added via script.
Simply calling ListenForEffectChanges in the Refresh function instead of
separately.
The reason for it being called above the text null check is that we can
add effects even if the text is null.